refactored setup file. NFACT now gets the lookup space file and coords file
