+++ NFACT/setup/configure_setup.py
@@ -3,7 +3,6 @@ from NFACT.utils.utils import (
     error_and_exit,
     read_file_to_list,
     colours,
-    make_directory,
     load_json,
 )
 from NFACT.pipes.image_handling import check_files_are_imaging_files
@@ -43,29 +42,6 @@ def get_subjects(args: dict) -> dict:
         return args
 
 
-def check_subject_exist(list_of_subjects: str) -> None:
-    """
-    Function to check that subjects.
-    Exit and errors if they don't
-
-    Parameters
-    ----------
-    list_of_subjects: str
-        list of subjects
-
-    Returns
-    -------
-    None
-    """
-    [
-        error_and_exit(
-            os.path.exists(path),
-            f"{path} does not exist. Please check provided subjects",
-        )
-        for path in list_of_subjects
-    ]
-
-
 def does_list_of_subjects_exist(path_to_list: str) -> bool:
     """
     Function to check if list of subjects
@@ -87,6 +63,29 @@ def does_list_of_subjects_exist(path_to_list: str) -> bool:
     if (not os.path.exists(path_to_list)) or (os.path.isdir(path_to_list)):
         return False
     return True
+
+
+def check_subject_exist(list_of_subjects: str) -> None:
+    """
+    Function to check that subjects.
+    Exit and errors if they don't
+
+    Parameters
+    ----------
+    list_of_subjects: str
+        list of subjects
+
+    Returns
+    -------
+    None
+    """
+    [
+        error_and_exit(
+            os.path.exists(path),
+            f"{path} does not exist. Please check provided subjects",
+        )
+        for path in list_of_subjects
+    ]
 
 
 def return_list_of_subjects_from_file(path_to_list: str) -> list:
@@ -142,84 +141,6 @@ def process_seeds(seeds: str) -> list:
 
     [check_files_are_imaging_files(seed) for seed in list_of_seeds]
     return list_of_seeds
-
-
-def create_folder_set_up(directory: str) -> None:
-    """
-    Function to create NFACT directory set up
-
-    Parameters
-    ----------
-    directory: str
-        path to directory to save output
-        in.
-
-    Returns
-    -------
-    None
-    """
-    error_and_exit(
-        os.path.exists(directory),
-        "Output directory does not exist. Please provide actual directory",
-    )
-    col = colours()
-    print(f"{col['purple']}nfact folder is in {directory}{col['reset']}")
-    nfact_directory = os.path.join(directory, "nfact")
-
-    sub_folders = [
-        "group_averages",
-        "ICA",
-        "NFM",
-        "GLM",
-        "ICA/dual_reg",
-        "NFM/dual_reg",
-        "ICA/normalised",
-        "NFM/normalised",
-        "ICA/dual_reg/G",
-        "ICA/dual_reg/W",
-        "NFM/dual_reg/G",
-        "NFM/dual_reg/W",
-    ]
-
-    does_exist = os.path.exists(nfact_directory)
-
-    if does_exist:
-        sub_folders = which_nfact_folders_exist(nfact_directory, sub_folders)
-
-        if len(sub_folders) == 0:
-            return None
-
-    if not does_exist:
-        make_directory(nfact_directory)
-
-    [make_directory(os.path.join(nfact_directory, sub)) for sub in sub_folders]
-
-
-def which_nfact_folders_exist(nfact_directory: str, sub_folders: list) -> list:
-    """
-    Function to check which nfact sub folders
-    don't exist
-
-    Parameters
-    ----------
-    nfact_directory: str
-        path to nfact directory
-    sub_folders: list
-        list of sub folders
-
-    Returns
-    -------
-    sub_folders_that_dont_exist: list
-        list of sub folders that don't exist
-    """
-
-    sub_folders_that_dont_exist = []
-    [
-        sub_folders_that_dont_exist.append(sub)
-        for sub in sub_folders
-        if not os.path.exists(os.path.join(nfact_directory, sub))
-    ]
-    return sub_folders_that_dont_exist
 
 
 def list_of_fdt_mat(list_ptx_folder: list) -> list:
